+++ src/dune_console/dune_console.ml
@@ -24,8 +24,49 @@ module Backend = struct
   let progress_no_flush = Progress.no_flush
 end
 
+(* Flag that controls whether messages should be separated by a blank line *)
+let separate_messages_flag = ref false
+
+(* A user message that solely contains a blank line *)
+let blank_line_msg =
+  { User_message.paragraphs = [ Pp.cut ]
+  ; hints = []
+  ; annots = User_message.Annots.empty
+  ; loc = None
+  }
+
+(** Prints a blank line *)
+let print_blank_line () =
+  let (module M : Backend_intf.S) = !Backend.main in
+  M.print_user_message blank_line_msg
+
+let first_msg = ref true
+
+let separate_messages v = separate_messages_flag := v
+
+(* If the [separate_messages = false], then [print_blank_line ()] does nothing.
+    When [separate_messages = true], [print_blank_line ()] does nothing the
+    first time it is called, whereas subsequent calls print a new line. Note
+    that calls to [reset] or [reset_flush_history] will erase the information
+    of whether some message has already been printed. As a consequence, after a
+    call to [reset] or [reset_flush_history], [print_blank_line] will behave as
+    if it has never been called before. *)
+let print_blank_line () =
+  if !separate_messages_flag then
+    (* only do something when the flag is on, i.e. the first time
+       the function is called *)
+    if !first_msg then
+      (* do not print anything the first time the function is
+         called, but remember it has been called at least once *)
+      first_msg := false
+    else
+      (* if the function has already been called at least once,
+         print a blank line *)
+      print_blank_line ()
+
 let print_user_message msg =
   let (module M : Backend_intf.S) = !Backend.main in
+  print_blank_line ();
   M.print_user_message msg
 
 let print paragraphs = print_user_message (User_message.make paragraphs)
@@ -41,10 +82,14 @@ let print_if_no_status_line line =
   M.print_if_no_status_line line
 
 let reset () =
+  (* forget that [print_user_message] has ever been called *)
+  first_msg := true;
   let (module M : Backend_intf.S) = !Backend.main in
   M.reset ()
 
 let reset_flush_history () =
+  (* forget that [print_user_message] has ever been called *)
+  first_msg := true;
   let (module M : Backend_intf.S) = !Backend.main in
   M.reset_flush_history ()