+++ CHANGES.md
@@ -4,6 +4,9 @@ Unreleased
 - Fix preludes not being recorded as dependencies in the `(mdx)` stanza (#7109,
   fixes #7077, @emillon).
 
+- Invoke preprocessor commands from directory of dune file containing the
+  commands rather than from the workspace root (#7057, fixes #7043, @gridbugs)
+
 3.7.0 (2023-02-17)
 ------------------